asdasd

+++ Data/Scripts/WeaponCore/WeaponComp/WeaponPlatform/WeaponTracking.cs
@@ -147,7 +147,7 @@ namespace WeaponCore.Platform
             weapon.IsTracking = inRange && newDesiredAz >= weapon.MinAzimuthRadians && newDesiredAz <= weapon.MaxAzimuthRadians && newDesiredEl >= weapon.MinElevationRadians && newDesiredEl <= weapon.MaxElevationRadians;
 
             if (!step) return weapon.IsTracking;
-            if (weapon.IsTracking && maxAzimuthStep > float.MinValue)
+            if (weapon.IsTracking && maxAzimuthStep > double.MinValue)
             {
                 var oldAz = weapon.Azimuth;
                 var oldEl = weapon.Elevation;
